+++ lib/util/util.py
@@ -4,6 +4,11 @@ import json
 import logging
 import re
 from packaging import version
+import urllib.parse
+
+from enodo.jobs import (
+    JOB_TYPE_DETECT_ANOMALIES_FOR_SERIES, JOB_TYPE_FORECAST_SERIES,
+    JOB_TYPE_STATIC_RULES)
 
 MIN_DATA_FILE_VERSION = "1.0.0"
 CURRENT_DATA_FILE_VERSION = "1.0.0"
@@ -29,6 +34,25 @@ def regex_valid(regex):
         return True
     except re.error:
         return False
+
+
+def parse_output_series_name(series_name, output_series_name):
+    # enodo_{series_name}_forecast_{job_config_name}
+    type_and_config_name = output_series_name.replace(
+        f"enodo_{series_name}_", "")
+    job_type = type_and_config_name.split("_")[0]
+    config_name = type_and_config_name.replace(f"{job_type}_", "")
+    return job_type, config_name
+
+
+def get_job_config_output_series_name(
+        series_name, job_type, config_name):
+    job_prefix_mapping = {
+        JOB_TYPE_FORECAST_SERIES: "forecast",
+        JOB_TYPE_DETECT_ANOMALIES_FOR_SERIES: "anomalies",
+        JOB_TYPE_STATIC_RULES: "static_rules"
+    }
+    return f"enodo_{series_name}_{job_prefix_mapping[job_type]}_{config_name}"
 
 
 def load_disk_data(path):
@@ -79,3 +103,90 @@ def cls_lock():
                 raise Exception("Incorrect usage of cls_lock func")
         return wrapped
     return wrapper
+
+
+def apply_fields_filter(resources: list, fields: list) -> list:
+    resources = [dict(t) for t in resources]
+    if fields is not None:
+        resources = [
+            filter_single_dict(resource, fields)
+            for resource in resources]
+    return resources
+
+
+def filter_single_dict(
+        resource: dict, fields: list) -> dict:
+    return {k: v for k, v in resource.items() if k in fields}
+
+
+def async_simple_fields_filter(
+        is_list=True, is_wrapped=True, return_index=None):
+    def wrapper(func):
+        @functools.wraps(func)
+        async def wrapped(*args, **kwargs):
+            fields = kwargs.pop('fields', None)
+            resp = await func(*args, **kwargs)
+            if return_index is not None:
+                org_resp = list(resp)
+                resp = resp[return_index]
+            if is_wrapped:
+                if 'data' not in resp:
+                    return resp  # might be an error
+                resp = resp.get('data', [] if is_list else {})
+            if fields is not None:
+                if is_list is True:
+                    resp = apply_fields_filter(resp, fields)
+                else:
+                    resp = filter_single_dict(resp, fields)
+            if is_wrapped:
+                resp = {"data": resp}
+            if return_index is not None:
+                org_resp[return_index] = resp
+                resp = tuple(org_resp)
+            return resp
+        return wrapped
+    return wrapper
+
+
+def sync_simple_fields_filter(
+        is_list=True, is_wrapped=True, return_index=None):
+    def wrapper(func):
+        @functools.wraps(func)
+        def wrapped(*args, **kwargs):
+            fields = kwargs.pop('fields', None)
+            resp = func(*args, **kwargs)
+            if return_index is not None:
+                org_resp = list(resp)
+                resp = resp[return_index]
+            if is_wrapped:
+                if 'data' not in resp:
+                    return resp  # might be an error
+                resp = resp.get('data', [] if is_list else {})
+            if fields is not None:
+                if is_list is True:
+                    resp = apply_fields_filter(resp, fields)
+                else:
+                    resp = filter_single_dict(resp, fields)
+            if is_wrapped:
+                resp = {"data": resp}
+            if return_index is not None:
+                org_resp[return_index] = resp
+                resp = tuple(org_resp)
+            return resp
+        return wrapped
+    return wrapper
+
+
+def implement_fields_query(func):
+    @functools.wraps(func)
+    async def wrapped(cls, request):
+        fields = None
+        if "fields" in request.rel_url.query:
+            fields = urllib.parse.unquote(
+                request.rel_url.query['fields'])
+            if fields == "":
+                fields = None
+            if fields is not None:
+                fields = fields.split(",")
+        return await func(cls, request, fields=fields)
+    return wrapped