+++ CHANGELOG.md
@@ -10,6 +10,31 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Resolve series per job to re-trigger
 
+
+## [0.1.0-beta3.2.20] - 2022-07-14
+
+### Removed
+- Removed old job config template routes and handlers
+
+### Added
+- Analysis region to series state job meta
+- Job meta to output
+
+### Changed
+- Implemented generic fields filter for multiple GET endpoints
+- Merged output routes and handlers into one with query param to filter by job_type
+
+## [0.1.0-beta3.2.19] - 2022-07-12
+
+### Fixed
+- Fixed bug with filtering fields for all output endpoint
+
+## [0.1.0-beta3.2.18] - 2022-07-12
+
+### Added
+- Fields filter function, applied to series output `ALL` GET endpoint
+- Return `job_config_name` and `job_type` when returning all series output
+
 ## [0.1.0-beta3.2.17] - 2022-07-07
 
 ### Added